SWATCH-1191 Remove swatch tally JMX endpoint deleteDataAssociatedWithOrg
