+++ src/test/java/org/candlepin/subscriptions/tally/admin/InternalTallyResourceTest.java
@@ -30,6 +30,8 @@ import javax.ws.rs.BadRequestException;
 import org.candlepin.subscriptions.ApplicationProperties;
 import org.candlepin.subscriptions.FixedClockConfiguration;
 import org.candlepin.subscriptions.retention.TallyRetentionController;
+import org.candlepin.subscriptions.security.SecurityProperties;
+import org.candlepin.subscriptions.tally.AccountResetService;
 import org.candlepin.subscriptions.tally.MarketplaceResendTallyController;
 import org.candlepin.subscriptions.tally.TallySnapshotController;
 import org.candlepin.subscriptions.tally.billing.RemittanceController;
@@ -50,10 +52,12 @@ class InternalTallyResourceTest {
   @Mock private TallySnapshotController snapshotController;
   @Mock private CaptureSnapshotsTaskManager snapshotTaskManager;
   @Mock private TallyRetentionController tallyRetentionController;
+  @Mock private AccountResetService accountResetService;
 
   private InternalTallyResource resource;
   private ApplicationProperties appProps;
   private ApplicationClock clock;
+  SecurityProperties properties;
 
   @BeforeEach
   void setupTest() {
@@ -67,7 +71,9 @@ class InternalTallyResourceTest {
             remittanceController,
             snapshotController,
             snapshotTaskManager,
-            tallyRetentionController);
+            tallyRetentionController,
+            accountResetService,
+            properties);
   }
 
   @Test
